backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic[email]
python-jose[cryptography]
python-dotenv
aiohttp
